+++ lib/cache/models/cache_limiter_base.dart
@@ -1,4 +1,0 @@
-abstract class CacheLimiterBase {
-  Future<bool> isLimited();
-  Future<void> update();
-}