+++ lib/core/guards/rate_limiter_guard.dart
@@ -1,0 +1,18 @@
+import 'package:firefast/core/guards/operation_limiter_base.dart';
+import 'package:firefast/firefast_core.dart';
+import 'package:prf/prf.dart';
+
+class RateLimiterGuard implements OperationLimiterBase, OperationGuard {
+  final PrfRateLimiter limiter;
+
+  const RateLimiterGuard(this.limiter);
+
+  @override
+  Future<bool> isLimited() async => await limiter.isLimitedNow();
+
+  @override
+  Future<void> update() async => await limiter.tryConsume();
+
+  @override
+  Future<bool> canPerform() async => await isLimited();
+}